Pulling intro art to log file
