+++ packages/build-tools/utils/log.js
@@ -1,5 +1,7 @@
 const chalk = require('chalk');
 
+// @todo Can we have emojis? How does `yarn` handle it?
+
 /**
  * Logs a successful message to Terminal
  * @param {string} msg
@@ -71,6 +73,47 @@ Task Done: ${task}
 ===`));
 }
 
+/**
+ * Displays big Bolt intro ASCII art
+ */
+function intro() {
+// @TODO: How best to pull in .scss color data export to reference branch colors below?
+  const b = chalk.hex('#545DA6'); // indigo, light
+  const y = chalk.hex('#ffcc4d'); // yellow, base
+
+  const CLI_TITLE = chalk.bold.underline('Bolt-CLI');
+  const CLI_DESCRIPTION = 'Welcome to the Bolt CLI ⚡️  Have fun!';
+  const CLI_USAGE = 'Usage: \`bolt <command> [options ...]\`';
+
+// const HELP_HEADER_BACKUP = `
+//     /˜˜˜˜˜˜˜˜˜˜˜˜\
+//    / /˜˜˜˜/|˜˜˜˜\ \
+//   / /    / |     \ \
+//  / /    /  |____  \ \
+// / /    /       /   \ \
+// \ \   /____   /    / /
+//  \ \      |  /    / /
+//   \ \     | /    / /
+//    \ \____|/____/ /
+//     \____________/
+// `
+
+// ^ Colorized + partially filled in version of logo above
+  const HELP_HEADER = `
+    ${b('///////|\\\\\\\\\\\\')}
+   ${b('///˜˜˜˜')}${y('/|')}${b('˜˜˜˜\\\\\\')}
+  ${b('///    ')}${y('//|')}${b('     \\\\\\')}     ${CLI_TITLE}
+ ${b('///    ')}${y('///|____')}${b('  \\\\\\')}
+${b('///    ')}${y('/////////')}${b('   \\\\\\')}   ${CLI_DESCRIPTION}
+${b('\\\\\\   ')}${y('/////////')}${b('    ///')}
+ ${b('\\\\\\      ')}${y('|///')}${b('    ///')}    ${CLI_USAGE}
+  ${b('\\\\\\     ')}${y('|//')}${b('    ///')}
+   ${b('\\\\\\____')}${y('|/')}${b('____///')}
+    ${b('\\\\\\\\\\\\|///////')}
+`;
+
+  info(HELP_HEADER);
+}
 
 module.exports = {
   success,
@@ -80,4 +123,5 @@ module.exports = {
   errorAndExit,
   taskStart,
   taskDone,
+  intro,
 };
